refactor!: Improve DebugPrimitiveRenderer structure
- The `DebugPrimitiveRenderer` class has been refactored to enhance its structure and functionality.
